% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{applyQCScoreModel}
\alias{applyQCScoreModel}
\title{applyQCScoreModel}
\usage{
applyQCScoreModel(spe, qcModel, scoreName = "QC_score")
}
\arguments{
\item{spe}{A `SpatialExperiment` object with QC metrics already computed.}

\item{qcModel}{A QC score model object, usually stored in
`metadata(spe)$QCScore_model`.}

\item{scoreName}{Name of the output column in `colData`.}
}
\value{
A `SpatialExperiment` object with added QC score in `colData`.
}
\description{
Apply a previously trained QC score model to a new SpatialExperiment object.
}
\examples{

example(readCosmxSPE)

## Compute per-cell quality control metrics
spe <- spatialPerCellQC(spe)

## Split the object only for example purposes
set.seed(1998)
idx <- sample(seq_len(ncol(spe)), floor(ncol(spe) / 2))
spe_train <- spe[, idx]
spe_test <- spe[, -idx]

## Train the Quality Control (QC) score model on one dataset
spe_train <- computeQCScore(spe_train)
qc_model <- metadata(spe_train)$QCScore_model

## Apply the trained model to another dataset
spe_test <- applyQCScoreModel(
    spe=spe_test,
    qcModel=qc_model,
    scoreName="QC_score_transferred"
)

summary(spe_test$QC_score_transferred)
}
