% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{computeAreaFromPolygons}
\alias{computeAreaFromPolygons}
\title{computeAreaFromPolygons}
\usage{
computeAreaFromPolygons(polygons, coldata)
}
\arguments{
\item{polygons}{An `sf` object containing polygon data.}

\item{coldata}{A `DataFrame` containing the `colData` to which area information will be added.}
}
\value{
A `DataFrame` with the added area information.
}
\description{
This function computes the area from polygon data and adds it to the `colData`.
}
\examples{
# Assuming `polygons` is an sf object and `coldata` is a DataFrame:
# coldata <- computeAreaFromPolygons(polygons, coldata)
}
