% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeBorderDistanceCosMx}
\alias{computeBorderDistanceCosMx}
\title{computeBorderDistanceCosMx}
\usage{
computeBorderDistanceCosMx(
  spe,
  xwindim = metadata(spe)$fov_dim[["xdim"]],
  ywindim = metadata(spe)$fov_dim[["ydim"]]
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial transcriptomics data.}

\item{xwindim}{Width of the FOV in the x-dimension. Defaults to
`metadata(spe)$fov_dim[["xdim"]]`.}

\item{ywindim}{Height of the FOV in the y-dimension. Defaults to
`metadata(spe)$fov_dim[["ydim"]]`.}
}
\value{
The `SpatialExperiment` object with added border distance data in
`colData`.
}
\description{
Compute Distance to FoV Border in SpatialExperiment for CosMx technology.

Calculates the minimum distance of each coordinate in a `SpatialExperiment`
object to the nearest border of the field of view (FOV) and adds it to
`colData`.
}
