% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeFilterFlags}
\alias{computeFilterFlags}
\title{computeFilterFlags}
\usage{
computeFilterFlags(
  spe,
  fs_threshold = 0.5,
  use_fs_quantiles = FALSE,
  total_threshold = 0,
  ctrl_tot_ratio_threshold = 0.1
)
}
\arguments{
\item{spe}{a SpatialExperiment object with total probe counts, control probe
counts on total counts ratio and flag score in the `colData`,
tipically computed with \link{spatialPerCellQC}}

\item{ctrl_tot_ratio_threshold}{}
}
\description{
defines flagged cells for variables based on the thresholds passed as input
}
