% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeFixedFlags}
\alias{computeFixedFlags}
\title{computeFixedFlags}
\usage{
computeFixedFlags(spe, total_threshold = 0, ctrl_tot_ratio_threshold = 0.1)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial transcriptomics data.}

\item{total_threshold}{A numeric value for the threshold of total counts to
identify cells with zero counts. Default is `0`.}

\item{ctrl_tot_ratio_threshold}{A numeric value for the threshold of
control-to-total ratio to flag outliers. Default is `0.1`.}
}
\value{
The `SpatialExperiment` object with added filter flags in `colData`.
}
\description{
Compute Flagged cells using fixed thresholds for SpatialExperiment.

This function calculates various flags to identify outliers in a
`SpatialExperiment` object based on quality control metrics.
}
\details{
The function flags cells based on zero counts, control-to-total
ratio, and `quality_score` to identify potential outliers. It also combines
these flags into a single filter flag.
}
\examples{
example(readCosmxSPE)
spe <- spatialPerCellQC(spe)
spe <- computeFixedFlags(spe)
table(spe$fixed_filter_out)
}
