% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXenium.R
\name{computeMissingMetricsXenium}
\alias{computeMissingMetricsXenium}
\title{computeMissingMetricsXenium}
\usage{
computeMissingMetricsXenium(pol_file, coldata, keep_polygons = FALSE)
}
\arguments{
\item{pol_file}{A character string specifying the file path to the polygon
data.}

\item{coldata}{A `DataFrame` containing the `colData` for the Xenium dataset.}

\item{keep_polygons}{A logical value indicating whether to keep the polygon
data in the resulting `colData`. Default is `FALSE`.}
}
\value{
A `DataFrame` containing the updated `colData` with computed metrics.
If `keep_polygons` is `TRUE`, the polygon data is also included.
}
\description{
Compute Missing Metrics for Xenium Data

This function computes missing metrics, such as the aspect ratio, from
polygon data in a Xenium dataset and optionally appends the polygon data to
the resulting `colData`.
}
\details{
The function reads the polygon data from the specified file,
computes the aspect ratio for each polygon, and merges these metrics with
the provided `colData`. Optionally, the polygon data can be kept in the
returned `colData`.
}
\examples{
# Assuming 'pol_file' is the path to the polygon file and 'coldata' is a
# DataFrame:
#updated_cd <- computeMissingMetricsXenium(pol_file = "path/to/polygons",
#                                          coldata = coldata,
#                                          keep_polygons = TRUE)
}
