% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeQCScore}
\alias{computeQCScore}
\title{computeQCScore (deprecated)}
\usage{
computeQCScore(spe, bestLambda = NULL, modelFormula = NULL, verbose = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object.}

\item{bestLambda}{Passed to \code{\link{computeQScore}}.}

\item{modelFormula}{Passed to \code{\link{computeQScore}}.}

\item{verbose}{Passed to \code{\link{computeQScore}}.}
}
\value{
A `SpatialExperiment` object; see \code{\link{computeQScore}}.
}
\description{
\strong{Deprecated.} Use \code{\link{computeQScore}} instead.
}
