% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeQCScoreFlags}
\alias{computeQCScoreFlags}
\title{computeQCScoreFlags (deprecated)}
\usage{
computeQCScoreFlags(spe, qsThreshold = 0.5, useQSQuantiles = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object.}

\item{qsThreshold}{Passed to \code{\link{computeQScoreFlags}}.}

\item{useQSQuantiles}{Passed to \code{\link{computeQScoreFlags}}.}
}
\value{
A `SpatialExperiment` object; see \code{\link{computeQScoreFlags}}.
}
\description{
\strong{Deprecated.} Use \code{\link{computeQScoreFlags}} instead.
}
