% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeQCScoreFlags}
\alias{computeQCScoreFlags}
\title{computeQCScoreFlags}
\usage{
computeQCScoreFlags(spe, qs_threshold = 0.5, use_qs_quantiles = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial transcriptomics data.}

\item{qs_threshold}{Numeric threshold or quantile for QC score. Default
`0.5`.}

\item{use_qs_quantiles}{Logical; if `TRUE`, treat `qs_threshold` as a
percentile.}
}
\value{
The `SpatialExperiment` object with added filter flags in `colData`.
}
\description{
Compute flagged cells based on a manually chosen threshold on QC score

This function Compute flagged cells based on a manually chosen threshold on
QC score stored in `SpatialExperiment` object.
}
\examples{
example(computeQCScore)
spe <- computeQCScoreFlags(spe)
table(spe$is_qcscore_outlier)
# if fixed filters are defined we have an additional column
spe <- computeThresholdFlags(spe)
spe <- computeQCScoreFlags(spe)
table(spe$threshold_qcscore_flags)
}
