% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeQScoreFlags}
\alias{computeQScoreFlags}
\title{computeQScoreFlags}
\usage{
computeQScoreFlags(spe, qsThreshold = 0.5, useQSQuantiles = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial omics data.}

\item{qsThreshold}{Numeric threshold or quantile for QS. Default is `0.5`.}

\item{useQSQuantiles}{Logical; if `TRUE`, treat `qsThreshold` as a percentile.}
}
\value{
The `SpatialExperiment` object with added low-quality flagged cells as
TRUE/FALSE in `colData`.
}
\description{
Computes flagged cells based on a manually chosen threshold on Quality Score (QS).

This function computes flagged cells based on a manually chosen threshold on
QS stored in `SpatialExperiment` object.
}
\examples{
example(computeQScore)
spe <- computeQScoreFlags(spe)
table(spe$low_QScore)
# if fixed filters are defined we have an additional column
spe <- computeThresholdFlags(spe)
spe <- computeQScoreFlags(spe)
table(spe$low_threshold_QScore)
}
