% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{customPolyMetrics}
\alias{customPolyMetrics}
\title{customPolyMetrics}
\usage{
customPolyMetrics(spe = spe)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with custom polygons stored inside after
running addPolygonsToSpe.}
}
\value{
A `SpatialExperiment` with new metrics
}
\description{
This function computes centroids, area, area in um, aspect ratio
and logged target counts on area ratio for custom polygons. New variables have
a prefix cust_ to distinguish them from the previous variables.
}
\examples{
#TBD
}
