% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{.addPolygonsToCD}
\alias{.addPolygonsToCD}
\title{Attach sf polygons to a DataFrame of cell metadata}
\usage{
.addPolygonsToCD(cd, polygons, polygonsCol = "polygons")
}
\arguments{
\item{cd}{A DataFrame containing at least `fov` and `cellID` columns.}

\item{polygons}{An sf object with matching `fov` and `cellID` columns.}

\item{polygonsCol}{character indicating the name of the polygons column to
add into the colData (default is `polygons`).}
}
\value{
A DataFrame identical to `cd`, but row‑subset to cells present in
`polygons` and with a new `polygons` list‑column of sf geometries.
}
\description{
This function enriches a DataFrame (e.g., from colData) with matching polygon
geometries.
}
\keyword{internal}
