% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{.checkPolygonsValidity}
\alias{.checkPolygonsValidity}
\title{Title}
\usage{
.checkPolygonsValidity(sf)
}
\arguments{
\item{sf}{}
}
\description{
checks validity on the active geometry of `sf` parameter
}
\keyword{internal}
