% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{.checkSkw}
\alias{.checkSkw}
\title{.checkSkw}
\usage{
.checkSkw(spe = spe, metric_list = metric_list)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial omics data.}

\item{metric_list}{A character vector specifying the metrics to include in
the QC score formula. Default is `c("log2CountArea", "Area_um",
"log2AspectRatio", "log2Ctrl_total_ratio")`.}
}
\value{
A vector containing the list of chosen outlier detection method for each
metric.
}
\description{
Check skewness of metrics to choose outlier detection method for
`SpatialExperiment`.
}
\examples{
example(readCosmxSPE)
spe <- spatialPerCellQC(spe)
.checkSkw(spe, metric_list = c("log2CountArea", "Area_um", "log2AspectRatio",
"log2Ctrl_total_ratio")`.

}
