% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{.computeBorderDistanceCosMx}
\alias{.computeBorderDistanceCosMx}
\title{.computeBorderDistanceCosMx}
\usage{
.computeBorderDistanceCosMx(
  spe,
  xwindim = metadata(spe)$fov_dim[["xdim"]],
  ywindim = metadata(spe)$fov_dim[["ydim"]]
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with CosMx data.}

\item{xwindim}{Width of FOV along x axis (default from `metadata(spe)$fov_dim`).}

\item{ywindim}{Height of FOV along y axis (default from `metadata(spe)$fov_dim`).}
}
\value{
A `SpatialExperiment` object with `dist_border` column in
`colData`.
}
\description{
Calculates the minimum distance of each cell to the Field‐Of‐View (FOV) 
border and adds it to `colData`.
}
\keyword{internal}
