% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{dot-computeCosmxTrainSet}
\alias{dot-computeCosmxTrainSet}
\alias{.computeCosmxTrainSet}
\title{.computeCosmxTrainSet}
\usage{
.computeCosmxTrainSet(spe)
}
\arguments{
\item{spe}{\code{SpatialExperiment}}
}
\value{
A list with elements \code{bad} and \code{good}, each a data.frame
with \code{qscore_train} and (for “good”) an \code{is_a_bad_boy} flag.
}
\description{
Internal: builds Training Set for CosMx.
Splits a SpatialExperiment into “bad” vs “good” cells based on
outliers in aspect ratio near FOV border or low signal density.
}
\keyword{internal}
