% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{dot-computeXenMerTrainSet}
\alias{dot-computeXenMerTrainSet}
\alias{.computeXenMerTrainSet}
\title{.computeXenMerTrainSet}
\usage{
.computeXenMerTrainSet(spe)
}
\arguments{
\item{spe}{\code{SpatialExperiment}}
}
\value{
A list with elements \code{bad} and \code{good}, each a data.frame
with \code{qscore_train} and (for “good”) an \code{is_a_bad_boy} flag.
}
\description{
Internal: builds Training Set for Xenium & MERFISH.
Splits a SpatialExperiment into “bad” vs “good” cells based on
pre-computed outlier labels on log2SignalDensity.
}
\keyword{internal}
