% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{.createPolygons}
\alias{.createPolygons}
\title{.createPolygons}
\usage{
.createPolygons(
  spat_obj,
  x = NULL,
  y = NULL,
  polygon_id = NULL,
  geometry = "Geometry"
)
}
\arguments{
\item{polygon_id}{}
}
\description{
.createPolygons
}
\keyword{internal}
