% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{.dark_theme}
\alias{.dark_theme}
\title{.dark_theme}
\usage{
.dark_theme(fill_color = "black", fore_color = "white")
}
\arguments{
\item{fore_color}{color for all the other elements (default is "white")}
}
\value{
a ggplot2 theme object
}
\description{
internal function to setup the black background theme for the First Filter
plot
#' @param fill_color color to fill the element_rect (default is "black")
}
\keyword{internal}
