% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{.fov_image_theme}
\alias{.fov_image_theme}
\title{.fov_image_theme}
\usage{
.fov_image_theme(back.color = "black", back.border = NA, title.col = "white")
}
\arguments{
\item{back.color}{not used}

\item{back.border}{color for the borders of the background (default=NA)}

\item{title.col}{character indicating the color of the title}
}
\value{
a ggplot2 theme object
}
\description{
internal function to setup the theme for the fov background on the whole
image
}
\keyword{internal}
