% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getActiveGeometryName}
\alias{.getActiveGeometryName}
\title{.getActiveGeometryName}
\usage{
.getActiveGeometryName(sf)
}
\arguments{
\item{sf}{an sf object}
}
\value{
character with the name of the active geometry
}
\description{
.getActiveGeometryName
}
