% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{.getAttrThr}
\alias{.getAttrThr}
\title{.getAttrThr}
\usage{
.getAttrThr(spe, attribute, threshold = c("lower", "higher", "both"))
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{attribute}{character indicating the colData columns of outlier.filter
class which the threshold has to be retrieved from.}

\item{threshold}{character, one of "lower"/"higher".}
}
\value{
a numeric value named as the `threshold`.
}
\description{
.getAttrThr
}
\examples{
example(computeSpatialOutlier)
.getAttrThr(spe, "log2CountArea_outlier_sc", "lower")
.getAttrThr(spe, "log2CountArea_outlier_sc", "higher")
.getAttrThr(spe, "log2CountArea_outlier_sc", "both")
}
\keyword{internal}
