% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{.getModelFormula}
\alias{.getModelFormula}
\title{Internal: Model Formula Printer}
\usage{
.getModelFormula(technology)
}
\arguments{
\item{technology}{\[character\]
Technology name to decide which predictors to include.}
}
\value{
\[character\]
  A one‐sided formula as a string (e.g. "~ log2CountArea + ...").
}
\description{
Returns the right‐hand side of a model formula string based on technology.
}
\keyword{internal}
