% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{.light_theme}
\alias{.light_theme}
\title{.light_theme}
\usage{
.light_theme(fill_color = "white", fore_color = "black")
}
\value{
a ggplot2 theme object
}
\description{
internal function to setup the white background theme for the First Filter plot
}
\keyword{internal}
