% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{.negative_image_theme}
\alias{.negative_image_theme}
\title{.negative_image_theme}
\usage{
.negative_image_theme(fill_color = "black", fore_color = "white")
}
\arguments{
\item{fill_color}{color to fill the element_rect (default is "black")}

\item{fore_color}{color for all the other elements (default is "white")}
}
\value{
a ggplot2 theme object
}
\description{
internal function to setup the theme for the negative background for
negative plots
}
\keyword{internal}
