% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.renameGeometry}
\alias{.renameGeometry}
\title{.renameGeometry}
\usage{
.renameGeometry(sf, from, to, activate = FALSE)
}
\arguments{
\item{sf}{an sf object with the `from` geometry}

\item{from}{character indicating the name of the geometry to change}

\item{to}{character indicating the new name of the geometry}

\item{activate}{logical indicating if the renamed geometry has to be activated}
}
\value{
an sf object
}
\description{
renames the `from` to `to` geometry of the `sf` object.
If `activate` is `TRUE` it set as the active geometry the new geometry name.
Default behaviour is to check if the renamed geometry is already active and
leave it as active with the new name.
}
