% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.setActiveGeometry}
\alias{.setActiveGeometry}
\title{.setActiveGeometry}
\usage{
.setActiveGeometry(sf, name)
}
\arguments{
\item{sf}{an sf object}

\item{name}{character for the geometry to activate}
}
\value{
an sf object
}
\description{
.setActiveGeometry
}
