% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{.setAttrThr}
\alias{.setAttrThr}
\title{.setAttrThr}
\usage{
.setAttrThr(spe, attribute, value)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{attribute}{character indicating the colData columns of `outlier.filter`
class which the threshold has to be assigned to.}

\item{value}{a numeric numeric value named with one of "lower"/"higher".
It sets the threshold for the `outlier.filter`.}
}
\value{
an updated SpatialExperiment object.
}
\description{
.setAttrThr
}
\examples{
example(computeSpatialOutlier)
thr <- .getAttrThr(spe, "log2CountArea_outlier_sc", "lower")
spe <- .setAttrThr(spe, "log2CountArea_outlier_sc", thr)
spe$log2CountArea_outlier_sc
}
\keyword{internal}
