% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\docType{data}
\name{firstFlagPalette}
\alias{firstFlagPalette}
\title{firstFlagPalette}
\format{
An object of class \code{character} of length 6.
}
\usage{
firstFlagPalette
}
\value{
a palette for firstFlagPlot
}
\description{
neon color palette for firstFlagPlot
}
\keyword{internal}
