% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialQCPlots.R
\name{plotCentroidsSpe}
\alias{plotCentroidsSpe}
\title{plotCentroidsSpe}
\usage{
plotCentroidsSpe(
  spe,
  colour_by = NULL,
  order_by = NULL,
  sample_id = unique(spe$sample_id),
  isNegativeProbe = FALSE,
  point_col = "darkmagenta",
  size = 0.05,
  alpha = 0.2,
  aspect_ratio = 1
)
}
\arguments{
\item{alpha}{}
}
\description{
plotCentroidsSpe
}
