% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialQCPlots.R
\name{plotPolygonsSPE_ggplot}
\alias{plotPolygonsSPE_ggplot}
\title{plotPolygonsSPE_ggplot}
\usage{
plotPolygonsSPE_ggplot(
  spe,
  colour_by = NULL,
  sample_id = unique(spe$sample_id),
  fill_alpha = 1,
  palette = NULL,
  border_col = NA,
  border_alpha = 1,
  border_line_width = 0.1,
  draw_borders = TRUE
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with polygon data as an `sf` object.}

\item{colour_by}{A column in `colData(spe)` for coloring the polygons.}

\item{sample_id}{Sample ID for plot title. Default is the unique sample ID.}

\item{fill_alpha}{Transparency level for polygon fill. Default is `1`.}

\item{palette}{Colors to use if `colour_by` is a factor. Default is `NULL`.}

\item{border_col}{Color of polygon borders. Default is `"black"`.}

\item{border_alpha}{Transparency level for borders. Default is `1`.}

\item{border_line_width}{Width of polygon borders. Default is `0.1`.}

\item{draw_borders}{Logical; whether to draw borders. Default is `TRUE`.}
}
\value{
A `ggplot` object representing the polygon plot of the spatial data.
}
\description{
Plot polygons from a `SpatialExperiment` object using ggplot2.
}
\examples{
# Assuming `spe` is a SpatialExperiment object with polygon data:
# plotPolygonsSPE_ggplot(spe, colour_by="gene_expression")
}
