% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCosMx.R
\name{readCosMx}
\alias{readCosMx}
\title{readCosMx}
\usage{
readCosMx(
  sample,
  path,
  rm.unmpd.tx = TRUE,
  fov = NULL,
  discard.tx = NULL,
  scalefactor = 0.18,
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sample}{character indicating the name of the sample}

\item{path}{character indicating the path pointing to the sample}

\item{rm.unmpd.tx}{logical (default TRUE) indicating if the unmapped
molecules have to be discarded}

\item{fov}{optional numeric indicating a list of FoVs to consider,
if NULL (default) all FoVs are considered}

\item{discard.tx}{character indicating a list of molecules to discard,
if NULL (default) all genes are kept}

\item{force}{logical, if TRUE the function uses the `path` as passed from
the user}

\item{verbose}{}
}
\description{
readCosMx
}
