% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{readh5polygons}
\alias{readh5polygons}
\title{readh5polygons}
\usage{
readh5polygons(polFile)
}
\arguments{
\item{polFile}{A character string specifying the file path to the HDF5
polygon data.}
}
\value{
A list containing the polygon geometries and their associated cell_id
}
\description{
This function reads polygon data from an HDF5 file.
}
\author{
Lambda Moses
}
\keyword{internal}
