% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMerfish.R
\name{readMerfishSPE}
\alias{readMerfishSPE}
\title{readMerfishSPE}
\usage{
readMerfishSPE(
  dirname,
  countmatfpattern = "cell_by_gene.csv",
  metadatafpattern = "cell_metadata.csv",
  polygonsfpattern = "cell_boundaries.parquet",
  coord_names = c("center_x", "center_y")
)
}
\arguments{
\item{fov_dims}{}
}
\value{
A SpatialExperiment object
}
\description{
readMerfishSPE
}
