% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{readPolygonsCosmx}
\alias{readPolygonsCosmx}
\title{readPolygonsCosmx}
\usage{
readPolygonsCosmx(
  polygonsFile,
  type = c("csv", "parquet"),
  x = "x_global_px",
  y = "y_global_px",
  xloc = "x_local_px",
  yloc = "y_local_px",
  keepMultiPol = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
readPolygonsCosmx
}
