% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatvectgenerate.R
\name{spatVectGenerate}
\alias{spatVectGenerate}
\title{spatVectGenerate}
\usage{
spatVectGenerate(
  main.path,
  image.type = "CellLabels",
  fov.number,
  flip.vertical = FALSE,
  force = FALSE
)
}
\arguments{
\item{main.path}{character for the main path of the CosMx sample}

\item{image.type}{character for the images folder name (default is
`CellLabels`)}

\item{fov.number}{numeric for the FoV id to process}

\item{flip.vertical}{logical indicating if the image has to be vertically
flipped}

\item{force}{logical, when TRUE it computes the fov polygons even if the file
already exists, it also overwrites it.}
}
\value{
a terra polygon object or an sf object in case of reading the object
}
\description{
Generates/loads the shape file for each FoV image present in the
`main.path\image.type`.
See details for additional information.
}
\details{
Generates/loads the shape file for each FoV image present in the
`main.path\image.type`.
In case the shape file already exists, the function loads the already
existing shape file, returning the loaded object.
If the `force` param is set to `TRUE` it computes the polygons even when the
file already exists and also overwrites it.
}
