% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{applyQCScoreModel}
\alias{applyQCScoreModel}
\title{applyQCScoreModel}
\usage{
applyQCScoreModel(spe, qcModel, scoreName = "QC_score")
}
\arguments{
\item{spe}{A `SpatialExperiment` object with QC metrics already computed.}

\item{qcModel}{A QC score model object, usually stored in
`metadata(spe)$QCScore_model`.}

\item{scoreName}{Name of the output column in `colData`.}
}
\value{
A `SpatialExperiment` object with added QC score in `colData`.
}
\description{
Apply a previously trained QC score model to a new SpatialExperiment object.
See details for important considerations when applying a model to a
different dataset.
}
\details{
The authors do not reccommend applying a QC score model trained on one
dataset to a different dataset, but this could be useful in some cases,
for example to transfer a model trained on a dataset core and the applied to
other cores from the same experiment.
The QC score model should have been trained on a dataset with the same
set of QC metrics and similar data distribution to have better
predictions. The function will check for the presence of required model
variables in the new dataset and will handle missing or extra variables
accordingly.
}
\examples{

example(readCosmxSPE)

## Compute per-cell quality control metrics
spe <- spatialPerCellQC(spe)

## Split the object only for example purposes
set.seed(1998)
idx <- sample(seq_len(ncol(spe)), floor(ncol(spe) / 2))
spe_train <- spe[, idx]
spe_test <- spe[, -idx]

## Train the Quality Control (QC) score model on one dataset
spe_train <- computeQCScore(spe_train)
qc_model <- metadata(spe_train)$QCScore_model

## Apply the trained model to another dataset
spe_test <- applyQCScoreModel(
    spe=spe_test,
    qcModel=qc_model,
    scoreName="QC_score_transferred"
)

summary(spe_test$QC_score_transferred)
}
