% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{computeAreaFromPolygons}
\alias{computeAreaFromPolygons}
\title{computeAreaFromPolygons}
\usage{
computeAreaFromPolygons(polygons)
}
\arguments{
\item{polygons}{An `sf` object containing polygon data.}
}
\value{
A `numeric` vector with the area information.
}
\description{
This function computes the area from polygon data.
}
\examples{
# Assuming `polygons` is an sf object:
# coldata <- computeAreaFromPolygons(polygons, coldata)
}
