% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{computeAreaFromPolygons}
\alias{computeAreaFromPolygons}
\title{computeAreaFromPolygons}
\usage{
computeAreaFromPolygons(polygons, coldata)
}
\arguments{
\item{coldata}{}
}
\description{
computeAreaFromPolygons
}
