% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{computeAspectRatioFromPolygons}
\alias{computeAspectRatioFromPolygons}
\title{computeAspectRatioFromPolygons}
\usage{
computeAspectRatioFromPolygons(polygons, coldata)
}
\arguments{
\item{polygons}{An `sf` object containing polygon data.}
}
\value{
A `numeric` vector with the aspect ratio information.
}
\description{
This function computes the aspect ratio from polygon data and adds it to the `colData`.
}
\examples{
# Assuming `polygons` is an sf object:
# ar <- computeAspectRatioFromPolygons(polygons)
}
