% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{computeAspectRatioFromPolygons}
\alias{computeAspectRatioFromPolygons}
\title{computeAspectRatioFromPolygons}
\usage{
computeAspectRatioFromPolygons(polygons, coldata)
}
\arguments{
\item{coldata}{}
}
\description{
computeAspectRatioFromPolygons
}
