% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeBorderDistanceCosMx}
\alias{computeBorderDistanceCosMx}
\title{Compute distance to FOV border for CosMx}
\usage{
computeBorderDistanceCosMx(
  spe,
  xwindim = metadata(spe)$fov_dim[["xdim"]],
  ywindim = metadata(spe)$fov_dim[["ydim"]]
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with CosMx data.}

\item{xwindim}{Width of FOV in x (default from `metadata(spe)$fov_dim`).}

\item{ywindim}{Height of FOV in y (default from `metadata(spe)$fov_dim`).}
}
\value{
A `SpatialExperiment` object with `dist_border` columns in
  `colData`.
}
\description{
Calculates the minimum distance of each cell to the field‐of‐view border
and adds it to `colData`.
}
\examples{
example(readCosmxSPE)
spe <- computeBorderDistanceCosMx(spe)
}
