% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeLambda}
\alias{computeLambda}
\title{computeLambda}
\usage{
computeLambda(modelMatrix, trainDF)
}
\arguments{
\item{modelMatrix}{`matrix`
The design matrix built from the training data, typically via
`model.matrix(as.formula(model_formula), data=trainDF)`.}

\item{trainDF}{`data.frame`
A data frame for QS model training that must include:
  Predictor columns: All columns referenced in the formula returned by `getModelFormula()`.
  `QScore_train` A binary (0/1) response vector to be modeled.}
}
\value{
`numeric`
  The value of \eqn{\lambda} (i.e., `lambda.min`) from
  `cv.glmnet` that minimizes the cross-validation error.
}
\description{
Computes optimal ridge regularization parameter \eqn{\lambda} via
k-fold cross-validation.

\code{computeLambda} performs ridge (L2) logistic regression with
k-fold cross-validation to identify the optimal regularization parameter
\eqn{\lambda} for Quality Score (QS) model training.
}
\details{
Internally, the function:
  runs k-fold cross-validation of ridge logistic regression using `cv.glmnet` with `alpha = 0`,
  extracts and returns `ridge_cv$lambda.min`.
}
\examples{
example(computeTrainDF)
modform <- getModelFormula(names(metadata(spe)$formula_variables))
model_matrix <- model.matrix(as.formula(modform), data=df_train)
best_lambda <- computeLambda(model_matrix, df_train)
print(best_lambda)


}
