% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeLambda}
\alias{computeLambda}
\title{computeLambda}
\usage{
computeLambda(modelMatrix, trainDF)
}
\arguments{
\item{trainDF}{`data.frame`
A data frame for QS model training that must include:
  Predictor columns: All columns referenced in the formula returned by `getModelFormula()`.
  `qscore_train` A binary (0/1) response vector to be modeled.}

\item{modelFormula}{`character`
A character string representing the model formula
 `~ log2SignalDensity + ...`, as returned by `getModelFormula()`.}
}
\value{
`numeric`
  The value of \eqn{\lambda} (i.e., `lambda.min`) from
  `cv.glmnet` that minimizes the cross-validation error.
}
\description{
Computes optimal ridge regularization parameter \eqn{\lambda} via
k-fold cross-validation.

\code{computeLambda} performs ridge (L2) logistic regression with
k-fold cross-validation to identify the optimal regularization parameter
\eqn{\lambda} for Quality Score (QS) model training.
}
\details{
Internally, the function:
  constructs the design matrix via \code{model.matrix()},
  runs k-fold cross-validation of ridge logistic regression using `cv.glmnet` with `alpha = 0`,
  extracts and returns `ridge_cv$lambda.min`.
}
\examples{
example(computeTrainDF)
modform <- getModelFormula(metadata(spe)$formula_variables)
best_lambda <- computeLambda(df_train, modform)
print(best_lambda)


}
