% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeLambda}
\alias{computeLambda}
\title{computeLambda}
\usage{
computeLambda(spe, train_df, model_formula)
}
\arguments{
\item{spe}{\linkS4class{SpatialExperiment}
A SpatialExperiment containing at least:
\itemize{
  \item assay(s) with nonzero \code{total} counts,
  \item \code{colData(spe)} columns including \code{log2CountArea},
  \code{Area_um}, \code{log2Ctrl_total_ratio}, etc.
}}

\item{train_df}{\[data.frame\]
A data frame for training that must include:
\describe{
  \item{Predictor columns}{All columns referenced in the formula returned
  by \code{getModelFormula()}.}
  \item{\code{qcscore_train}}{A binary (0/1) response vector to be modeled.}
}}
}
\value{
\[numeric\]
  The value of \eqn{\lambda} (i.e., \code{lambda.min}) from
  \code{\link[glmnet]{cv.glmnet}} that minimizes the cross-validation error.
}
\description{
Compute Optimal Ridge Regularization Parameter \eqn{\lambda} via
Cross-Validation

\code{computeLambda} performs ridge (L2) logistic regression with
cross-validation to identify the optimal regularization parameter
\eqn{\lambda} for a binary response.
}
\details{
Internally, the function:
\enumerate{
  \item Constructs the design matrix via \code{model.matrix()},
  \item Runs ridge logistic regression cross-validation using
        \code{\link[glmnet]{cv.glmnet}} with \code{alpha = 0},
  \item Extracts and returns \code{ridge_cv$lambda.min}.
}
}
\examples{
example(spatialPerCellQC)
withr::with_seed(1998, train_df <- computeTrainDF(spe))
best_lambda <- computeLambda(spe, train_df, model_formula)
print(best_lambda)

}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
