% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeQScore}
\alias{computeQScore}
\title{computeQScore}
\usage{
computeQScore(spe, best_lambda = NULL, verbose = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial transcriptomics data.}

\item{best_lambda}{the best lambda typically computed using `computeLambda`.}

\item{verbose}{logical for having a verbose output. Default is FALSE.}
}
\value{
The `SpatialExperiment` object with added quality score in `colData`.
}
\description{
Compute quality score and automatically define weights for quality score
through glm training. This function computes quality score with a formula
that depends on the technology.
}
\details{
For CosMx datasets, the Quality Score formula is defined as follows:

quality score ~ count density - aspect ratio - interaction term

count density is total counts-to-area ratio, aspect ratio represents
border effect typical of CosMx datasets and the last one is the
interaction term of the previous two terms.

For Xenium and Merscope datasets, quality score depends solely on count
density, as no border effect has been observed for these two technologies.

To automatically define the formula coefficient weights, model training
is performed through ridge regression.
}
\examples{
example(spatialPerCellQC)
set.seed(1998)
spe <- computeQScore(spe)
summary(spe$training_status)
summary(spe$quality_score)
}
