% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeQScore}
\alias{computeQScore}
\title{computeQScore}
\usage{
computeQScore(spe, verbose = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial transcriptomics data.}

\item{verbose}{logical for having a verbose output. Default is FALSE.}
}
\value{
The `SpatialExperiment` object with added quality score in `colData`.
}
\description{
Compute quality score and automatically define weights for quality score
through glm training. This function computes quality score with a formula
that depends on the technology.
To automatically define the formula coefficient weights, model training
is performed through ridge regression.
}
\examples{
example(spatialPerCellQC)
set.seed(1998)
spe <- computeQScore(spe)
summary(spe$training_status)
summary(spe$quality_score)
}
