% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeQscoreFlags}
\alias{computeQscoreFlags}
\title{computeQscoreFlags}
\usage{
computeQscoreFlags(spe, qs_threshold = 0.5, use_qs_quantiles = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial transcriptomics data.}

\item{qs_threshold}{Numeric threshold or quantile for quality score. Default
`0.5`.}

\item{use_qs_quantiles}{Logical; if `TRUE`, treat `qs_threshold` as a
percentile.}
}
\value{
The `SpatialExperiment` object with added filter flags in `colData`.
}
\description{
Compute flagged cells based on a manually chosen threshold on quality score

This function Compute flagged cells based on a manually chosen threshold on
quality score stored in `SpatialExperiment` object.
}
\examples{
example(computeQScore)
spe <- computeQscoreFlags(spe)
table(spe$fixed_qscore_out)
}
