% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeQscoreFlags}
\alias{computeQscoreFlags}
\title{computeQscoreFlags}
\usage{
computeQscoreFlags(spe, qs_threshold = 0.5, use_qs_quantiles = FALSE)
}
\arguments{
\item{spe}{A `SpatialExperiment` object with spatial transcriptomics data.}

\item{use_qs_quantiles}{a boolean. If TRUE uses the value specified in
qs_threshold as a percentile}

\item{opt}{a boolean value to set to TRUE if you want to compute flagged cells
for optimized quality score}
}
\value{
The `SpatialExperiment` object with added filter flags in `colData`.
}
\description{
Compute flagged cells based on a manually chosen threshold on quality score

This function Compute flagged cells based on a manually chosen threshold on
quality score stored in `SpatialExperiment` object.
}
\examples{
# TBD

}
