% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeTrainDF}
\alias{computeTrainDF}
\title{computeTrainDF}
\usage{
computeTrainDF(colData, formulaVars, tech, verbose = FALSE)
}
\arguments{
\item{verbose}{\[logical(1)\] (default \code{FALSE})
If \code{TRUE}, prints the number of “bad” and “good” cells selected.}

\item{spe}{\linkS4class{SpatialExperiment}
A SpatialExperiment containing at least:
\itemize{
  \item assay(s) with nonzero \code{total} counts,
  \item \code{colData(spe)} columns including \code{log2CountArea},
  \code{Area_um}, \code{log2Ctrl_total_ratio}, etc.
}}
}
\value{
A \code{data.frame} with one row per cell, including:
\itemize{
  \item \code{qcscore_train} (0/1) indicating “bad” vs “good”,
  \item relevant \code{colData} columns used for modeling.
  \item Deduplicates and down-samples “good” cells to match the number of
   “bad” cells.
}
}
\description{
Build a Balanced Training Data Frame from a SpatialExperiment

\code{computeTrainDF} takes a \linkS4class{SpatialExperiment} object
and assembles a balanced training set of “good” vs “bad” cells for
subsequent model fitting.
}
\details{
The function builds a training set using the variables specified
in the `metadata` of the `SpatialExperiment` object.
}
\examples{
example(spatialPerCellQC)
spe <- .computeOutliersQCScore(spe)
spe <- .checkOutliers(spe)
df_train <- computeTrainDF(colData(spe), metadata(spe)$formula_variables,
    metadata(spe)$technology)
table(df_train$qcscore_train)

}
