% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXenium.R
\name{dot-addFovFromTx}
\alias{dot-addFovFromTx}
\alias{.addFovFromTx}
\title{.addFovFromTx}
\usage{
.addFovFromTx(txFile, colData)
}
\arguments{
\item{colData}{A `DataFrame` containing the `colData` for the Xenium dataset.}

\item{dirName}{`character(1)`
Path to a Xenium Output Bundle directory.}

\item{txPattern}{A character string specifying the pattern to look for in the
directory to find the transcript file, only parquet file is supported.
data.}
}
\value{
A `DataFrame` containing the updated `colData` with FOV information.
}
\description{
Add FOV information from transcript file to cell metadata.

This function retrieves FOV information from transcript file and appends
the data to the resulting `colData`.
}
\details{
The function reads the transcript file then groups it by cell_id
and merges the FOV information to the cell metadata in `colData`. Only
parquet file is supported for this operation
}
\keyword{internal}
