% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{.centroid_image_theme}
\alias{.centroid_image_theme}
\title{.centroid_image_theme}
\usage{
.centroid_image_theme(backBorder = NA)
}
\arguments{
\item{backBorder}{color for the borders of the background (default=NA)}
}
\value{
a ggplot2 theme object
}
\description{
internal function to setup the theme for the centroid plot background
}
\keyword{internal}
