% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{dot-checkSkw}
\alias{dot-checkSkw}
\alias{.checkSkw}
\title{.checkSkw}
\usage{
.checkSkw(
  cd,
  metricList = c("log2SignalDensity", "Area_um", "log2AspectRatio",
    "log2Ctrl_total_ratio")
)
}
\arguments{
\item{cd}{colData of `SpatialExperiment` object.}

\item{metricList}{A character vector specifying the metrics to include in
the QC score formula. Defaults are "log2SignalDensity", "Area_um",
"log2AspectRatio", "log2Ctrl_total_ratio".}
}
\value{
A vector containing the list of chosen outlier detection method for each
metric.
}
\description{
Check skewness of metrics to choose outlier detection method.
}
\keyword{internal}
