% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{.checkSkw}
\alias{.checkSkw}
\title{.checkSkw}
\usage{
.checkSkw(
  cd,
  metricList = c("log2CountArea", "Area_um", "log2AspectRatio", "log2Ctrl_total_ratio")
)
}
\arguments{
\item{cd}{colData of `SpatialExperiment` object.}

\item{metricList}{A character vector specifying the metrics to include in
the QC score formula. Default is `c("log2CountArea", "Area_um",
"log2AspectRatio", "log2Ctrl_total_ratio")`.}
}
\value{
A vector containing the list of chosen outlier detection method for each
metric.
}
\description{
Check skewness of metrics to choose outlier detection method.
}
\examples{
example(readCosmxSPE)
.checkSkw(colData(spe))`.

}
\keyword{internal}
