% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{dot-computeCosmxProteinTrainSet}
\alias{dot-computeCosmxProteinTrainSet}
\alias{.computeCosmxProteinTrainSet}
\title{.computeCosmxProteinTrainSet}
\usage{
.computeCosmxProteinTrainSet(spe)
}
\arguments{
\item{spe}{\code{SpatialExperiment}}
}
\value{
A list with elements \code{bad} and \code{good}, each a data.frame
with \code{qscore_train} and (for “good”) an \code{is_a_bad_boy} flag.
}
\description{
Internal: Build Training Set for CosMx-Protein
Splits a SpatialExperiment into “bad” vs “good” cells based on
outliers in aspect ratio near tissue border or low count area.
}
\keyword{internal}
