% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{.computeCosmxTrainSet}
\alias{.computeCosmxTrainSet}
\title{Internal: Build Training Set for CosMx}
\usage{
.computeCosmxTrainSet(spe)
}
\arguments{
\item{spe}{\linkS4class{SpatialExperiment}}
}
\value{
A list with elements \code{bad} and \code{good}, each a data.frame
with \code{qscore_train} and (for “good”) an \code{is_a_bad_boy} flag.
}
\description{
Splits a SpatialExperiment into “bad” vs “good” cells based on
outliers in aspect ratio near tissue border or low count area.
}
\keyword{internal}
