% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{.createPolygons}
\alias{.createPolygons}
\title{.createPolygons}
\usage{
.createPolygons(
  spat_obj,
  x = NULL,
  y = NULL,
  polygon_id = NULL,
  geometry = "Geometry"
)
}
\arguments{
\item{spat_obj}{A data frame or similar object containing spatial data.}

\item{x}{A character vector specifying the x-coordinates.}

\item{y}{A character vector specifying the y-coordinates.}

\item{polygon_id}{A character string specifying the polygon ID.}
}
\value{
An `sf` object containing the created polygons.
}
\description{
This internal function creates polygons from a spatial data object.
}
\keyword{internal}
