% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{.fov_image_theme}
\alias{.fov_image_theme}
\title{.fov_image_theme}
\usage{
.fov_image_theme(backColor = "black", backBorder = NA, titleCol = "white")
}
\arguments{
\item{backColor}{not used}

\item{backBorder}{color for the borders of the background (default=NA)}

\item{titleCol}{character indicating the color of the title}
}
\value{
a ggplot2 theme object
}
\description{
internal function to setup the theme for the fov background on the whole
image
}
\keyword{internal}
