% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialQCPlots.R
\name{plotCellsFovs}
\alias{plotCellsFovs}
\title{plotCellsFovs}
\usage{
plotCellsFovs(
  spe,
  point_col = "darkmagenta",
  numbers_col = "black",
  alpha_numbers = 0.8,
  sample_id = NULL
)
}
\arguments{
\item{spe}{A SpatialExperiment object}

\item{point_col}{the color to use for the dots in the plot
(default is "darkmagenta")}

\item{sample_id}{string indicating the sample to plot (default is NULL,
implying a single sample in the spe)}
}
\value{
a ggplot2 object
}
\description{
Creates a scatter plot with all the centroids of the cells distributed inside
their own FoVs as an overlapping grid.
}
\examples{
TBD
}
