% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialQCPlots.R
\name{plotCentroidsSPE}
\alias{plotCentroidsSPE}
\title{plotCentroidsSPE}
\usage{
plotCentroidsSPE(
  spe,
  colour_by = NULL,
  order_by = NULL,
  sample_id = unique(spe$sample_id),
  isNegativeProbe = FALSE,
  palette = NULL,
  point_col = "darkmagenta",
  size = 0.05,
  alpha = 0.2,
  aspect_ratio = 1,
  legend_point_size = 2,
  legend_point_alpha = 0.8
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object containing spatial
transcriptomics data.}

\item{colour_by}{An optional character string specifying the column in
`colData(spe)` to use for coloring the points. If `NULL`, all points will be
colored the same.}

\item{order_by}{An optional character string specifying the column in
`colData(spe)` to use for ordering the points.}

\item{sample_id}{A character string specifying the sample identifier to be
used as the plot title. (Default is the unique sample ID from `spe`)}

\item{isNegativeProbe}{A logical value indicating whether to apply a custom
color gradient for negative probe data. (Default is `FALSE`)}

\item{palette}{A vector of colors to be used as a custom palette. For
categorical data, this should be a vector of colors with the same length as
the number of levels in `colour_by`. For continuous data, this should be a
vector of colors used to create a gradient.}

\item{point_col}{A character string specifying the color of the points when
`colour_by` is `NULL`. (Default is `"darkmagenta"`)}

\item{size}{A numeric value specifying the size of the points. (Default is
`0.05`)}

\item{alpha}{A numeric value specifying the transparency level of the points.
(Default is `0.2`)}

\item{aspect_ratio}{A numeric value specifying the aspect ratio of the plot.
(Default is `1`)}

\item{legend_point_size}{A numeric value specifying the size of the points
in the legend. (Default is `2`)}

\item{legend_point_alpha}{A numeric value specifying the transparency level
of the points in the legend. (Default is `0.8`)}
}
\value{
A `ggplot` object representing the spatial coordinates plot of
polygon centroids.
}
\description{
Plot Spatial Coordinates for a SpatialExperiment Object
This function generates a ggplot of spatial coordinates from a
`SpatialExperiment` object, optionally coloring the points by a specified
column in `colData`.
}
\examples{
#TBD
}
