% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialQCPlots.R
\name{plotPolygonsSPE}
\alias{plotPolygonsSPE}
\title{plotPolygonsSPE}
\usage{
plotPolygonsSPE(
  spe,
  colour_by = NULL,
  sample_id = unique(spe$sample_id),
  fill_alpha = NA,
  palette = NULL,
  border_col = NA,
  border_alpha = NA,
  border_line_width = 0.1
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object containing spatial transcriptomics
data, including polygon data as sf object.}

\item{colour_by}{An optional character string specifying the column in
`colData(spe)` to use for coloring the points. If `NULL`, all points will be
colored the same. (Defatult is `NULL`)}

\item{title}{A character string specifying the title of the plot.
(Default is the unique sample ID from `spe`)}
}
\value{
A `tmap` plot object displaying the polygons.
}
\description{
Plot Polygons from a SpatialExperiment Object
This function generates a plot of polygons stored in a `SpatialExperiment`
object.
}
\examples{
#TBD
}
