% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialQCPlots.R
\name{plotQCscoreTerms}
\alias{plotQCscoreTerms}
\title{plotQCScoreTerms}
\usage{
plotQCscoreTerms(
  spe,
  sample_id = unique(spe$sample_id),
  palette = NULL,
  size = 0.05,
  alpha = 0.2,
  aspect_ratio = 1,
  custom = FALSE
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object containing spatial transcriptomics
data.}

\item{sample_id}{Must match values in the `fov` column of `colData(spe)`.}

\item{custom}{A boolean value. If TRUE, custom polygons derived metrics will
be used.}
}
\value{
A panel with multiple plots showing quality score terms
}
\description{
Plots the terms combined together in the quality score formula to allow viewing
which term impacts most on the quality score.

This function Plots the terms combined together in the quality score formula
to allow viewing which term impacts most on the quality score within a
`SpatialExperiment` object.
}
\examples{
# Assuming 'spe' is a SpatialExperiment object with FOVs and polygon data:
# plotZoomFovsMap(spe, fovs = c("FOV1", "FOV2"), colour_by = "cell_type",
#                title = "Zoomed FOVs with Polygons")
}
