% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCosMx.R
\name{readCosmxSPE}
\alias{readCosmxSPE}
\title{readCosmxSPE}
\usage{
readCosmxSPE(
  dirname,
  sample_name = "sample01",
  coord_names = c("CenterX_global_px", "CenterY_global_px"),
  countmatfpattern = "exprMat_file.csv",
  metadatafpattern = "metadata_file.csv",
  polygonsfpattern = "polygons.csv",
  fovposfpattern = "fov_positions_file.csv",
  fov_dims = c(xdim = 4256, ydim = 4256)
)
}
\arguments{
\item{fov_dims}{}
}
\value{
A SpatialExperiment object
}
\description{
readCosmxSPE
}
