% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{readPolygons}
\alias{readPolygons}
\title{readPolygons}
\usage{
readPolygons(
  polygonsFile,
  type = c("csv", "parquet", "h5"),
  x = c("x_global_px", "vertex_x"),
  y = c("y_global_px", "vertex_y"),
  xloc = "x_local_px",
  yloc = "y_local_px",
  keepMultiPol = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{
an sf object with the loaded and validated polygons
}
\description{
readPolygons
}
