% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{readPolygonsCosMx}
\alias{readPolygonsCosMx}
\title{readPolygonsCosMx}
\usage{
readPolygonsCosMx(
  polygonsFile,
  x = "x_global_px",
  y = "y_global_px",
  xloc = "x_local_px",
  yloc = "y_local_px",
  micronConvFact = 0.12
)
}
\arguments{
\item{micronConvFact}{}
}
\description{
readPolygonsCosMx
}
