% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{readPolygonsMerfish}
\alias{readPolygonsMerfish}
\title{readPolygonsMerfish}
\usage{
readPolygonsMerfish(
  polygonsFolder,
  type = c("HDF5", "parquet"),
  keepMultiPol = TRUE,
  hdf5pattern = "hdf5",
  z_lev = 3L,
  zcolumn = "ZIndex",
  geometry = "Geometry",
  verbose = FALSE
)
}
\arguments{
\item{hdf5pattern}{}
}
\description{
readPolygonsMerfish
}
