% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{readPolygonsXenium}
\alias{readPolygonsXenium}
\title{Title}
\usage{
readPolygonsXenium(
  polygonsFile,
  type = c("parquet", "csv"),
  x = "vertex_x",
  y = "vertex_y",
  keepMultiPol = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
