% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXenium.R
\name{readXeniumSPE}
\alias{readXeniumSPE}
\title{Load data from a 10x Genomics Xenium experiment}
\usage{
readXeniumSPE(
  dirname,
  sample_name = "sample01",
  type = c("HDF5", "sparse"),
  coord_names = c("x_centroid", "y_centroid"),
  boundaries_type = c("parquet", "csv"),
  compute_missing_metrics = TRUE,
  keep_polygons = FALSE,
  countsfilepattern = "cell_feature_matrix",
  metadatafpattern = "cells",
  polygonsfpattern = "cell_boundaries",
  polygonsCol = "polygons"
)
}
\arguments{
\item{dirname}{`character(1)`
Path to a Xenium Output Bundle directory.}

\item{sample_name}{`character(1)`
Sample identifier to assign to `sample_id`. Default: `"sample01"`.}

\item{type}{`character(1)`
One of `"HDF5"` or `"sparse"`; method to read the feature matrix.}

\item{coord_names}{`character(2)`
Names of X/Y spatial coordinate columns. Default:
`c("x_centroid", "y_centroid")`.}

\item{boundaries_type}{`character(1)`
One of `"parquet"` or `"csv"`; format of the polygon file.}

\item{compute_missing_metrics}{`logical(1)`
If `TRUE`, compute area and aspect‐ratio from boundary polygons.}

\item{keep_polygons}{`logical(1)`
If `TRUE`, append raw polygon geometries to `colData`.}

\item{countsfilepattern}{`character(1)`
Pattern to locate the feature matrix file. Default:
`"cell_feature_matrix"`.}

\item{metadatafpattern}{`character(1)`
Pattern to locate the cell metadata file. Default: `"cells"`.}

\item{polygonsfpattern}{`character(1)`
Pattern to locate the cell boundaries file. Default:
`"cell_boundaries"`.}

\item{polygonsCol}{`character(1)`
Name of the polygons column to add to `colData`. Default:
`"polygons"`.}
}
\value{
A [`SpatialExperiment`] object with assays, `colData`, spatial
  coordinates, and `metadata$polygons` & `metadata$technology`.
}
\description{
Creates a [`SpatialExperiment`] from an unzipped Xenium Output Bundle
directory containing spatial gene expression data.
}
\details{
Expects the unzipped bundle to contain an `outs/` folder with:
- `cell_feature_matrix.h5` or `cell_feature_matrix/`
- `cells.csv.gz`
}
\examples{
xepath <- system.file(
  "extdata", "Xenium_small", package = "SpaceTrooper"
)
(xe_spe <- readXeniumSPE(
  dirname = xepath,
  keep_polygons = TRUE
))
}
\author{
Estella Yixing Dong, Dario Righelli
}
