% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{spatialPerCellQC}
\alias{spatialPerCellQC}
\title{Perform per‐cell QC on a SpatialExperiment}
\usage{
spatialPerCellQC(
  spe,
  micronConvFact = 0.12,
  negProbList = c("NegPrb", "Negative", "SystemControl", "Ms IgG1", "Rb IgG",
    "NegControlProbe", "NegControlCodeword", "UnassignedCodeword", "Blank")
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object containing spatial data.}

\item{micronConvFact}{Numeric factor to convert pixels to microns. Default
`0.12`.}

\item{negProbList}{Character vector of patterns to identify negative probes.
Defaults include:
- Nanostring CosMx: `"NegPrb"`, `"Negative"`, `"SystemControl"`
- Xenium: `"NegControlProbe"`, `"NegControlCodeword"`,
  `"UnassignedCodeword"`
- MERFISH: `"Blank"`}
}
\value{
A `SpatialExperiment` object with added QC metrics in `colData`.
}
\description{
Computes quality‐control metrics for each cell and adds them to `colData`.
}
\details{
Calculates sums and detected counts for control and target probes,
computes ratio and count‐area metrics, converts coords to microns for
CosMx, and drops zero‐count cells.
}
\examples{
example(readCosmxSPE)
spe <- spatialPerCellQC(spe)
}
