% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{spatialPerCellQC}
\alias{spatialPerCellQC}
\title{spatialPerCellQC}
\usage{
spatialPerCellQC(
  spe,
  negProbList = c("NegPrb", "Negative", "SystemControl", "NegControlProbe",
    "NegControlCodeWord", "UnassignedCodeWord", "Blank"),
  micronConvFact = 0.12
)
}
\arguments{
\item{micronConvFact}{}
}
\description{
spatialPerCellQC
}
\examples{
spe <- readCosmxSPE("~/Downloads/CosMx_data/DBKero/CosMx_Breast/CosMx_data_Case2/", sample_name="DBKero_BC")
spe <- spatialPerCellQC(spe)
}
