% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{spatialPerCellQC}
\alias{spatialPerCellQC}
\title{spatialPerCellQC}
\usage{
spatialPerCellQC(
  spe,
  micronConvFact = 0.12,
  negProbList = c("NegPrb", "Negative", "SystemControl", "NegControlProbe",
    "NegControlCodeWord", "UnassignedCodeWord", "Blank")
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object containing spatial transcriptomics
data.}

\item{micronConvFact}{A numeric value for converting pixel dimensions to
microns. Default is `0.12`.}

\item{negProbList}{A character vector of patterns used to identify negative
probes.
Default values are:
- For Nanostring CosMx: `"NegPrb"`, `"Negative"`, `"SystemControl"`
- For Xenium: `"NegControlProbe"`, `"NegControlCodeWord"`,
`"UnassignedCodeWord"`
- For MERFISH: `"Blank"`}
}
\value{
The `SpatialExperiment` object with added quality control metrics in
`colData`.
}
\description{
Perform Per-Cell Quality Control on a SpatialExperiment Object

This function calculates quality control metrics for each cell in a
`SpatialExperiment` object and adds them to `colData`.
}
\details{
The function computes several QC metrics, including control probe
sums, target probe sums, and the ratio of control probes to the total.
}
